Implements collectible coins and extra lives.
Adds functionality for collecting coins and extra lives within the game.

Introduces new entities for coins and extra lives, and updates the
game logic to handle player interaction with these objects. The `PlayerState` class is introduced to manage player-related data such as lives and credits.

Also refactors GameObject to use floats for width and height.

+++ app/src/main/java/hr/foi/rmai/memento/entities/GameObject.kt
@@ -8,8 +8,8 @@ import hr.foi.rmai.memento.utils.RectHitbox
 import hr.foi.rmai.memento.views.Animation
 
 abstract class GameObject(
-    val width: Int,
-    val height: Int,
+    val width: Float,
+    val height: Float,
     var animFrameCount: Int,
     val bitmapName: String,
     val type: Char
@@ -85,8 +85,8 @@ abstract class GameObject(
         )
         bitmap = Bitmap.createScaledBitmap(
             bitmap,
-            (width * animFrameCount * pixelsPerMeter),
-            (height * pixelsPerMeter),
+            (width * animFrameCount * pixelsPerMeter).toInt(),
+            (height * pixelsPerMeter).toInt(),
             false
         )
 
@@ -108,8 +108,8 @@ abstract class GameObject(
         anim = Animation(
             bitmapName,
             pixelsPerMeter,
-            width,
-            height,
+            width.toInt(),
+            height.toInt(),
             animFrameCount,
             animFps
         )
